@@ -1,30 +1,35 @@
 module github.com/gallir/smart-relayer
 
-go 1.16
+go 1.20
 
 require (
-	github.com/BurntSushi/toml v1.2.0
-	github.com/aws/aws-sdk-go v1.44.111
+	github.com/BurntSushi/toml v1.2.1
+	github.com/aws/aws-sdk-go v1.44.250
+	github.com/bytedance/sonic v1.8.8 // indirect
 	github.com/cespare/xxhash v1.1.0
-	github.com/fasthttp/router v1.4.12
+	github.com/fasthttp/router v1.4.18
 	github.com/gabrielperezs/streamspooler v1.0.0
 	github.com/gallir/bytebufferpool v1.0.1
-	github.com/gin-gonic/gin v1.8.1
-	github.com/go-playground/validator/v10 v10.11.1 // indirect
-	github.com/goccy/go-json v0.9.11 // indirect
-	github.com/golang/protobuf v1.5.2
+	github.com/gin-gonic/gin v1.9.0
+	github.com/go-playground/validator/v10 v10.12.0 // indirect
+	github.com/goccy/go-json v0.10.2 // indirect
+	github.com/golang/protobuf v1.5.3
 	github.com/golang/snappy v0.0.4
-	github.com/klauspost/compress v1.15.11
-	github.com/mattn/go-isatty v0.0.16 // indirect
+	github.com/klauspost/compress v1.16.5
+	github.com/klauspost/cpuid/v2 v2.2.4 // indirect
+	github.com/leodido/go-urn v1.2.3 // indirect
+	github.com/mattn/go-isatty v0.0.18 // indirect
 	github.com/modern-go/concurrent v0.0.0-20180306012644-bacd9c7ef1dd // indirect
-	github.com/pelletier/go-toml/v2 v2.0.5 // indirect
+	github.com/pelletier/go-toml/v2 v2.0.7 // indirect
 	github.com/pquerna/ffjson v0.0.0-20190930134022-aa0246cd15f7
 	github.com/spaolacci/murmur3 v1.1.0
+	github.com/ugorji/go/codec v1.2.11 // indirect
 	github.com/valyala/bytebufferpool v1.0.1-0.20180905182247-cdfbe9377474 // indirect
-	github.com/valyala/fasthttp v1.40.0
-	golang.org/x/crypto v0.0.0-20220926161630-eccd6366d1be // indirect
-	golang.org/x/net v0.0.0-20221004154528-8021a29435af // indirect
-	golang.org/x/sys v0.0.0-20220928140112-f11e5e49a4ec
+	github.com/valyala/fasthttp v1.46.0
+	golang.org/x/arch v0.3.0 // indirect
+	golang.org/x/crypto v0.8.0 // indirect
+	golang.org/x/sys v0.7.0
 	golang.org/x/xerrors v0.0.0-20200804184101-5ec99f83aff1 // indirect
-	google.golang.org/protobuf v1.28.1 // indirect
+	google.golang.org/protobuf v1.30.0 // indirect
+	gopkg.in/yaml.v2 v2.4.0 // indirect
 )
